Refactor to simplify and add comments.
